backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1
openai==1.82.0
chromadb-client==0.6.3
sentence-transformers==4.1.0
langdetect==1.0.9
python-multipart==0.0.20
websockets==15.0.1
httpx==0.28.1
beautifulsoup4==4.13.4
requests==2.32.3
pytest==8.3.5
pytest-asyncio==0.25.3
